rename `--bgv-add-client-interface` to `--lwe-add-client-interface`
