+++ lib/Dialect/BGV/CMakeLists.txt
@@ -1,8 +1,6 @@
 add_subdirectory(IR)
-add_subdirectory(Transforms)
 add_mlir_dialect_library(MLIRBGV
         IR/BGVDialect.cpp
-        Transforms/AddClientInterface.cpp
 
         ADDITIONAL_HEADER_DIRS
         ${CMAKE_CURRENT_SOURCE_DIR}/IR
@@ -10,7 +8,6 @@ add_mlir_dialect_library(MLIRBGV
         DEPENDS
         MLIRBGVIncGen
         MLIRBGVOpsIncGen
-        MLIRBGVPassesIncGen
 
         LINK_LIBS PUBLIC
         MLIRIR
