+++ tools/BUILD
@@ -46,8 +46,6 @@ cc_binary(
         "@heir//lib/Conversion/SecretToBGV",
         "@heir//lib/Dialect/ArithExt/IR:Dialect",
         "@heir//lib/Dialect/BGV/IR:Dialect",
-        "@heir//lib/Dialect/BGV/Transforms",
-        "@heir//lib/Dialect/BGV/Transforms:AddClientInterface",
         "@heir//lib/Dialect/CGGI/IR:Dialect",
         "@heir//lib/Dialect/CGGI/Transforms",
         "@heir//lib/Dialect/CKKS/IR:Dialect",
